imporve prefix mapping
some OpenMS tools have  parameters with the same name (different
sections). This creates problems in Galaxy.

workaround is to hardcode a unique prefix.

Before _ was used which now creates problems with the tool
linting (parameters are not supposed to start with _).

Hence we now use prefix of the form DEDUP_[0-9]+_DEDUP_